implement new stack handling in the treewalk interpreter, which will allow allocation of structs

+++ parsing.h
@@ -189,7 +189,7 @@ struct expr
     union
     {
         uint64_t integer_value;
-        double float_value;
+        float float_value;
         bool bool_value;
         const char *name;
         const char *string_value;
